// Frontend_React/src/features/course/course.types.ts
export enum CourseStatus {
  Draft = "Draft",
  Published = "Published",
  Pending = "Pending",
  Banned = "Banned",
  Rejected = "Rejected",
}

export enum CourseLevel {
  Beginner = "Beginner",
  Intermediate = "Intermediate",
  Advanced = "Advanced",
  All = "All",
}
// ...
export interface CourseDetailVm {
  id: string;

  title: string;
  description: string;
  summary: string;

  sectionCount: number;
  lectureCount: number;

  sections: SectionVm[];

  imageUrl: string;
  promoVideoUrl: string;

  price: number;
  discountPercentage: number;
  discountedPrice: number;

  durationInSeconds: number;

  status: CourseStatus;

  level: CourseLevel;

  learningOutcomes: string[];
  prerequisites: string[];

  categoryName: string;

  enrollmentCount: number;
  reviewCount: number;
  averageRating: number;

  instructorId: string;
  instructorName: string;
  instructorAvatarUrl: string | null;
  instructorProfessionalTitle: string | null;
  instructorAbout: string | null;
  instructorAverageRating: number;
  instructorReviewCount: number;
  instructorCourseCount: number;
  instructorStudentCount: number;
  retryCount: number;
  bannedReason: string | null;
  approvalHistories: CourseApprovalHistoryVm[];

  ratingDistribution: CourseRatingDistributionVm;

  createdAt: string;
  updatedAt: string;
  submittedAt: string | null;
  lastRejectedAt: string | null;
}

export interface SetCourseBannedStatusCommand {
  id: string;
  isBanned: boolean;
  bannedReason?: string;
}

export interface SetCourseBannedStatusResponse {
  message: string;
  data: {
    id: string;
    status: string;
    bannedReason?: string;
  };
}

export interface ModerateCourseCommand {
  id: string;
  isApproved: boolean;
  note: string;
}

export interface CourseApprovalHistoryVm {
  id: string;
  courseId: string;
  isApproved: boolean;
  note: string;
  createdAt: string;
  updatedAt: string;
}
// ...
export interface CourseRatingDistributionVm {
  star1: number;
  star2: number;
  star3: number;
  star4: number;
  star5: number;
}